requirements.txt
numpy==1.24.3
matplotlib==3.7.2
torch==2.0.1
torchvision==0.15.2
pyro-ppl==1.8.6
scikit-learn==1.3.0
pillow==10.0.0
tqdm==4.65.0
pandas==2.0.3
pytest==7.4.0 